аутентификация апи и СВЭГер

+++ back/app/views.py
@@ -1,21 +1,19 @@
-from django.shortcuts import render
 from rest_framework.response import Response
-from rest_framework.renderers import JSONRenderer
 from rest_framework.decorators import api_view
 from .models import Songs, Users, Musicians
-from .serializers import SongSerializer, UserSerializer, MusicianSerializer
-from django.http import HttpResponse, JsonResponse
+from .serializers import SongSerializer, UserSerializer, MusicianSerializer, SongUploadSerializer, SongEditSerializer
+from django.http import JsonResponse
 import json
 from django.views.decorators.csrf import csrf_exempt
 from django.conf import settings
-from .forms import DataForm
-import xml.etree.ElementTree as ET
 import os
 from django.core.exceptions import ValidationError
 from mutagen import File
 from datetime import datetime, timedelta
 from django.shortcuts import get_object_or_404
 from django.db.models import Q
+from drf_yasg.utils import swagger_auto_schema
+
 
 @csrf_exempt
 def set_theme(request):
@@ -34,8 +32,23 @@ def get_theme(request):
     theme = request.COOKIES.get('theme', 'false') 
     return JsonResponse({'theme': theme})
 
- 
-@api_view(['GET', 'POST'])
+def validate_token(request):
+    # Извлекаем токен из заголовка Authorization
+    auth_header = request.headers.get('Authorization')
+    if not auth_header or not auth_header.startswith('Bearer '):
+        return None, JsonResponse({"error": "Токен отсутствует или имеет неверный формат."}, status=401)
+
+    token = auth_header.split(' ')[1] 
+    try:
+        user = Users.objects.get(token=token)
+        return user, None  # Возвращаем пользователя, если токен найден
+    except Users.DoesNotExist:
+        return None, JsonResponse({"error": "Неверный токен."}, status=401)
+
+# =======================================================================================
+# нужные views
+
+@api_view(['GET'])
 def getAllSongs(request):
     songs = Songs.objects.select_related('musician').all();
 
@@ -46,8 +59,26 @@ def getAllSongs(request):
     
     return Response({'songs': serializer.data})
 
-@api_view(['POST', 'GET'])
+@api_view(['GET'])
+def search_songs(request):
+    query = request.GET.get('query', '') 
+
+    songs = Songs.objects.all()
+
+    if query:
+
+        songs = songs.filter(
+            Q(namesong__icontains=query) | Q(musician__musiciannickname__icontains=query)
+        )
+
+    # Сериализация результатов
+    serializer = SongSerializer(songs, many=True)
+    
+    return JsonResponse({"songs": serializer.data})
+
+@api_view(['POST'])
 def loginUser(request):
+    
     if request.method == 'POST':
         try:
             data = json.loads(request.body)
@@ -55,8 +86,9 @@ def loginUser(request):
             password = data.get('password')
             user = Users.objects.get(email=email)
             if user.password == password:
+                user.generate_token()
                 serializer = UserSerializer(user)
-                return Response({'user' : serializer.data, 'message' : 'correct'})
+                return Response({'user' : serializer.data, 'message' : 'correct', 'token' : user.token})
             else:
                 return Response({'message':'incorrect'}) 
         except:
@@ -64,7 +96,7 @@ def loginUser(request):
     else:
         return Response({'message':'badrequest'})
 
-@api_view(['POST', 'GET'])
+@api_view(['POST'])
 def getMusician(request):
     if request.method == 'POST':
         try:
@@ -80,7 +112,7 @@ def getMusician(request):
     else:
         return Response({'message':'badrequest'})
 
-@api_view(['POST', 'GET'])
+@api_view(['POST'])
 def getMusicianSongs(request):
     if request.method == 'POST':
         try:
@@ -99,67 +131,18 @@ def getMusicianSongs(request):
     else:
         return Response({'message':'badrequest'})    
 
-@api_view(['POST', 'GET'])
-def saveTextForm(request):
-    if request.method == "POST":
-        form = DataForm(request.POST)
-        
-        if form.is_valid():
-
-            data = form.cleaned_data
-
-            file_path = os.path.join(settings.MEDIA_ROOT, f"{data['song']}.json")
-            with open(file_path, 'w') as f:
-                json.dump(data, f)
-                
-            return JsonResponse({"message": "success"})
-        else:
-            return JsonResponse({"errors": form.errors}, status=400)    
-
-@api_view(['POST', 'GET'])
-def saveTextFile(request):
-    if request.method == 'POST' and request.FILES.get('file'):
-        uploaded_file = request.FILES['file']
-        try:
-            # Сохраняем файл на диск
-            file_path = os.path.join(settings.MEDIA_ROOT, f"{request.POST['song']}.json")
-            with open(file_path, 'wb+') as destination:
-                for chunk in uploaded_file.chunks():
-                    destination.write(chunk)
-
-            # Перемещаем указатель файла в начало
-            uploaded_file.seek(0)
-
-            # Проверка валидности JSON
-            try:
-                json.load(uploaded_file)
-            except json.JSONDecodeError:
-                os.remove(file_path)
-                return JsonResponse({"error": "Invalid JSON file"}, status=400)
-            
-            return JsonResponse({"message": "success"})
-
-        except ValidationError as e:
-            return JsonResponse({"error": str(e)}, status=400)
-
-def list_files(request):
-    folder_path = settings.MEDIA_ROOT
-    files = os.listdir(folder_path)
-
-    if not files:
-        return JsonResponse({"message": "No files found"}, status=404)
-
-    file_list = []
-    for file in files:
-        if file.endswith('.json'):
-            file_list.append(file)
-    
-    return JsonResponse({"files": file_list})
-
-
-@api_view(['POST', 'GET'])
+@swagger_auto_schema(
+    method='post',
+    request_body=SongUploadSerializer,  
+    responses={200: 'Song successfully added', 400: 'Invalid input'}
+)
+@api_view(['POST'])
 def addSong(request):
 
+    user, error_response = validate_token(request)
+    if error_response:
+        return error_response
+    
     if request.method == 'POST' and request.FILES.get('audiofile'):
         uploaded_file = request.FILES['audiofile']
 
@@ -195,23 +178,105 @@ def addSong(request):
         except ValidationError as e:
             return JsonResponse({"error": str(e)}, status=400)
 
-@api_view(['PUT'])
-def editSong(request):
+# @api_view(['PUT'])
+# def editSong(request):
 
-    if request.method == 'PUT':
+    # if request.method == 'PUT':
 
-        try:
+    #     try:
 
-            song = get_object_or_404(Songs, pk=request.POST.get('id'))
+    #         song = get_object_or_404(Songs, pk=request.POST.get('id'))
 
-            song.namesong = request.POST.get('newnamesong')
-            song.genre = request.POST.get('newgenre')
+    #         song.namesong = request.POST.get('newnamesong')
+    #         song.genre = request.POST.get('newgenre')
             
-            new_file_path = os.path.join(settings.MEDIA_ROOT, f"audio_files/{request.POST.get('newnamesong')}.mp3")
+    #         new_file_path = os.path.join(settings.MEDIA_ROOT, f"audio_files/{request.POST.get('newnamesong')}.mp3")
 
-            # Переименовываем файл, если название изменилось
+    #         # Переименовываем файл, если название изменилось
+    #         old_file_path = os.path.join(settings.MEDIA_ROOT, f"audio_files/{request.POST.get('oldnamesong')}.mp3")
+
+    #         if request.POST.get('oldnamesong') != request.POST.get('newnamesong') and os.path.exists(old_file_path):
+    #             os.rename(old_file_path, new_file_path)
+    #             song.audiofile = f"audio_files/{request.POST.get('newnamesong')}.mp3"
+            
+    #         song.save()           
+            
+    #         return JsonResponse({"message": "success"})
+
+    #     except ValidationError as e:
+    #         return JsonResponse({"error": str(e)}, status=400)
+
+# @api_view(['DELETE'])
+# def deleteSong(request):
+#     if request.method == 'DELETE':
+
+#         try:
+#             file_path = os.path.join(settings.MEDIA_ROOT, f"audio_files/{request.data.get('namesong')}.mp3")
+
+#             os.remove(file_path)
+
+#             song = Songs.objects.get(pk=request.data.get('id'))
+#             song.delete()
+            
+#             return JsonResponse({"message": "success"})
+
+#         except Songs.DoesNotExist:
+#             return JsonResponse({"error": "Song not found"}, status=404)
+        
+
+@csrf_exempt
+@swagger_auto_schema(
+    method='put',
+    request_body=SongEditSerializer,  
+    responses={200: 'Song successfully added', 400: 'Invalid input'}
+)
+@api_view(['GET', 'PUT', 'DELETE'])
+def song_view(request, id=None):
+    
+    user, error_response = validate_token(request)
+    if error_response:
+        return error_response
+
+    if request.method == 'POST':
+        if not request.FILES.get('audiofile'):
+            return JsonResponse({"error": "Файл не найден"}, status=400)
+
+        uploaded_file = request.FILES['audiofile']
+        try:
+            file_path = os.path.join(settings.MEDIA_ROOT, f"audio_files/{request.POST['namesong']}.mp3")
+            with open(file_path, 'wb+') as destination:
+                for chunk in uploaded_file.chunks():
+                    destination.write(chunk)
+
+            # Определяем длительность песни
+            audio = File(file_path)
+            if audio is None or not hasattr(audio.info, 'length'):
+                return JsonResponse({"error": "Невозможно определить продолжительность аудиофайла."}, status=400)
+
+            song_duration = timedelta(seconds=int(audio.info.length))
+            song = Songs.objects.create(
+                musician=get_object_or_404(Musicians, pk=request.POST.get('musician')),
+                namesong=request.POST['namesong'],
+                genre=request.POST['genre'],
+                audiofile=f"audio_files/{request.POST['namesong']}.mp3",
+                songpreview=None,
+                songrealeasedatee=datetime.today().strftime("%Y-%m-%d"),
+                songduration=song_duration,
+            )
+            return JsonResponse({"message": "success"})
+
+        except ValidationError as e:
+            return JsonResponse({"error": str(e)}, status=400)
+
+    elif request.method == 'PUT':
+        try:
+            song = get_object_or_404(Songs, pk=id)
+            song.namesong = request.POST.get('newnamesong', song.namesong)
+            song.genre = request.POST.get('newgenre', song.genre)
+
             old_file_path = os.path.join(settings.MEDIA_ROOT, f"audio_files/{request.POST.get('oldnamesong')}.mp3")
-
+            new_file_path = os.path.join(settings.MEDIA_ROOT, f"audio_files/{request.POST.get('newnamesong')}.mp3")
+            
             if request.POST.get('oldnamesong') != request.POST.get('newnamesong') and os.path.exists(old_file_path):
                 os.rename(old_file_path, new_file_path)
                 song.audiofile = f"audio_files/{request.POST.get('newnamesong')}.mp3"
@@ -222,39 +287,33 @@ def editSong(request):
 
         except ValidationError as e:
             return JsonResponse({"error": str(e)}, status=400)
-
-@api_view(['DELETE'])
-def deleteSong(request):
-    if request.method == 'DELETE':
-
+    
+    elif request.method == 'DELETE':
         try:
-            file_path = os.path.join(settings.MEDIA_ROOT, f"audio_files/{request.data.get('namesong')}.mp3")
+            song = get_object_or_404(Songs, pk=id)
+            file_path = os.path.join(settings.MEDIA_ROOT, song.audiofile)
 
-            os.remove(file_path)
+            if os.path.exists(file_path):
+                os.remove(file_path)
 
-            song = Songs.objects.get(pk=request.data.get('id'))
             song.delete()
-            
             return JsonResponse({"message": "success"})
 
         except Songs.DoesNotExist:
-            return JsonResponse({"error": "Song not found"}, status=404)
+            return JsonResponse({"error": "Песня не найдена"}, status=404)
+
+    # GET: Получение информации о песне
+    elif request.method == 'GET':
         
-@api_view(['GET'])
-def search_songs(request):
-    query = request.GET.get('query', '')  # Название песни или имя музыканта
+        try:
 
-    # Начинаем с получения всех песен
-    songs = Songs.objects.all()
+            song = Songs.objects.select_related('musician').get(pk=id)   
 
-    # Если задан запрос
-    if query:
-        # Используем Q объекты для поиска по обеим полям с логикой OR
-        songs = songs.filter(
-            Q(namesong__icontains=query) | Q(musician__musiciannickname__icontains=query)
-        )
-
-    # Сериализация результатов
-    serializer = SongSerializer(songs, many=True)
+            serializer = SongSerializer(song)
     
-    return JsonResponse({"songs": serializer.data})
+            return Response({'song': serializer.data})
+
+        except Songs.DoesNotExist:
+            return JsonResponse({"error": "Песня не найдена"}, status=404)
+
+    return JsonResponse({"error": "Метод не поддерживается"}, status=405)
